+++ src/PermGroups.jl
@@ -14,7 +14,7 @@ Group([(1,2),(2,3)])
 
 # PermGroups are iterators over their elements
 julia> collect(G)  
-6-element Array{Perm{Int64},1}:
+6-element Array{Perm{Int16},1}:
  (1,2)
  (1,3,2)
  ()
@@ -37,19 +37,19 @@ false
 
 # a list of points that no element of G fixes
 julia> base(G) 
-2-element Array{Int64,1}:
+2-element Array{Int16,1}:
  1
  2
 
 # the i-th element is the centralizer of base[1:i-1]
 julia> centralizers(G) 
-2-element Array{PermGroup{Int64},1}:
+2-element Array{PermGroup{Int16},1}:
  Group([(1,2),(2,3)])
  Group([(2,3)])
 
 # i-th element is transversal of centralizer[i] on base[i]
 julia> transversals(G)
-2-element Array{Dict{Int64,Perm{Int64}},1}:
+2-element Array{Dict{Int16,Perm{Int16}},1}:
  Dict(2 => (1,2),3 => (1,3,2),1 => ())
  Dict(2 => (),3 => (2,3))
 ```
